Optomizing file structure
