+++ client/src/components/profile/Blogs.tsx
@@ -66,7 +66,12 @@ const Blogs = ({ devId, mediumId }: UserProps): ReactElement => {
       <Grid container spacing={4}>
         {blogs.length === 0 ? (
           <Grid item xs={12}>
-            <Typography variant='h1' component='h2' fontSize={'1rem'}>
+            <Typography
+              variant='h1'
+              component='h2'
+              fontSize={'1rem'}
+              align='center'
+            >
               No Blogs Found
             </Typography>
           </Grid>
@@ -74,9 +79,10 @@ const Blogs = ({ devId, mediumId }: UserProps): ReactElement => {
           blogs.map((blog, idx) => (
             <Grid item key={idx} xs={12} sm={6} md={4}>
               <Card>
+                {/* <AspectRatio> */}
+                {/* <Link href={blog.url}>{`${blog.cover_image}`}</Link> */}
                 <CardMedia
                   component='img'
-                  height='140'
                   image={blog.cover_image}
                   alt={blog.title}
                 />
@@ -87,7 +93,7 @@ const Blogs = ({ devId, mediumId }: UserProps): ReactElement => {
                     fontSize={'2rem'}
                     align='center'
                   >
-                    <Link href={blog.url}>{blog.title}</Link>
+                    {blog.title}
                   </Typography>
                 </CardContent>
               </Card>
