+++ client/src/components/profile/UserInfo.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { UserProfile } from '../../../types';
+import { UserProfile } from '../../../../types';
 import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
 import TextField from '@mui/material/TextField';
